da03完成菜品和套餐管理功能

+++ sky-server/src/main/java/com/sky/service/impl/CategoryServiceImpl.java
@@ -2,13 +2,20 @@ package com.sky.service.impl;
 
 import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
+import com.sky.constant.MessageConstant;
+import com.sky.constant.StatusConstant;
 import com.sky.context.BaseContext;
 import com.sky.dto.CategoryDTO;
 import com.sky.dto.CategoryPageQueryDTO;
 import com.sky.entity.Category;
+import com.sky.entity.Employee;
+import com.sky.exception.DeletionNotAllowedException;
 import com.sky.mapper.CategoryMapper;
+import com.sky.mapper.DishMapper;
 import com.sky.result.PageResult;
 import com.sky.service.CategoryService;
+import com.sky.service.DishService;
+import com.sky.service.SetmealService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -20,6 +27,10 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Autowired
     private CategoryMapper categoryMapper;
+    @Autowired
+    private DishService dishService;
+    @Autowired
+    private SetmealService setmealService;
 
     /**
      * 分页查询分类信息
@@ -48,6 +59,19 @@ public class CategoryServiceImpl implements CategoryService {
      */
     @Override
     public void deleteById(Long id) {
+        Category category = categoryMapper.getById(id);
+        Integer status = category.getStatus();
+        if(status==StatusConstant.ENABLE){
+            throw new DeletionNotAllowedException(MessageConstant.DISH_ON_SALE);
+        }
+        Integer dishCount = dishService.count(id);
+        Integer setmealCount = setmealService.count(id);
+        if(dishCount>0){
+            throw new DeletionNotAllowedException(MessageConstant.CATEGORY_BE_RELATED_BY_DISH);
+        }
+        if(setmealCount>0){
+            throw new DeletionNotAllowedException(MessageConstant.CATEGORY_BE_RELATED_BY_SETMEAL);
+        }
         categoryMapper.deleteById(id);
     }
 
@@ -58,8 +82,6 @@ public class CategoryServiceImpl implements CategoryService {
      */
     @Override
     public void update(Category category) {
-        category.setUpdateTime(LocalDateTime.now());
-        category.setUpdateUser(BaseContext.get());
         categoryMapper.update(category);
     }
 
@@ -70,11 +92,17 @@ public class CategoryServiceImpl implements CategoryService {
      */
     @Override
     public void add(Category category) {
-        category.setCreateTime(LocalDateTime.now());
-        category.setUpdateTime(LocalDateTime.now());
-        category.setCreateUser(BaseContext.get());
-        category.setUpdateUser(BaseContext.get());
-        category.setStatus(1);
+        category.setStatus(StatusConstant.DISABLE);
         categoryMapper.add(category);
+    }
+
+    /**
+     * 根据类型获取分类信息
+     * @param type
+     * @return
+     */
+    @Override
+    public List<Category> listType(Integer type) {
+        return categoryMapper.listType(type);
     }
 }
